backend/pyproject.toml
[project]
name = "code-tutor-ai"
version = "1.0.0"
description = "AI-based Python Algorithm Learning Platform"
readme = "README.md"
requires-python = ">=3.11"

dependencies = [
    # Web Framework
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "python-multipart>=0.0.6",

    # Database
    "sqlalchemy>=2.0.25",
    "alembic>=1.13.1",
    "asyncpg>=0.29.0",
    "redis>=5.0.1",

    # Auth
    "bcrypt>=4.0.0",
    "python-jose[cryptography]>=3.3.0",

    # Validation
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "email-validator>=2.1.0",

    # DI & Utils
    "dependency-injector>=4.41.0",
    "structlog>=24.1.0",

    # ML/AI (optional, for later phases)
    # "torch>=2.1.0",
    # "transformers>=4.36.0",
    # "langchain>=0.1.0",
    # "faiss-cpu>=1.7.4",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "httpx>=0.26.0",
    "aiosqlite>=0.19.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
]

ml = [
    "torch>=2.1.0",
    "transformers>=4.36.0",
    "langchain>=0.1.0",
    "faiss-cpu>=1.7.4",
    "bitsandbytes>=0.41.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/code_tutor"]

[tool.ruff]
line-length = 88
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]
target-version = "py311"

[tool.ruff.isort]
known-first-party = ["code_tutor"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "e2e: End-to-end tests",
]
addopts = "-v --tb=short"

[tool.coverage.run]
source = ["src/code_tutor"]
branch = true
omit = ["*/tests/*", "*/__init__.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
]
fail_under = 70
